backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.10.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
browserbase==1.4.0
build==1.3.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.1.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.2.2
crewai==0.203.0
crewai-tools==0.76.0
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.7.0
google-api-core==2.26.0
google-auth==2.40.3
google-genai==1.43.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.10.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kombu==5.5.4
kubernetes==33.1.0
lance-namespace==0.0.18
lance-namespace-urllib3-client==0.0.18
lancedb==0.25.2
langchain==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.25
litellm==1.74.9
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20250506
pdfplumber==0.11.7
pdftotext==3.0.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.55.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.38.2
pylatexenc==2.10
PyMuPDF==1.26.3
pypdf==6.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.0.0
redis==6.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
stagehand==0.5.4
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@903944a411c35b8f7b2b51ada77a7c2a80e7fb88
triton==3.3.1
typer==0.16.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.13
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
yarl==1.20.1
youtube-transcript-api==1.2.2
zipp==3.23.0
zstandard==0.24.0
